fix: satisfy errorprone, CheckerFramework, and checkstyle
Errorprone (-PenableErrorprone, -Werror):
- @SuppressWarnings("JavaUtilDate") around the legacy
  java.util.Date.getTime() casts in Date/Time/Timestamp/Timestamptz/
  TimetzCodec.
- Int8Codec / OidCodec: explicit cast to double/float to silence the
  LongDoubleConversion / LongFloatConversion lint.
- ByteaCodec: pin String→bytea to US_ASCII (DefaultCharset).
- FallbackCodec: drop the unused bytesToHex helper.
- PgDatabaseMetaData: drop the unused columnTypeOid local.
- PgSQLInputBinary.parseCompositeData no longer takes the unused
  fields list.
- TypeInfoCache / PgConnection get the missing @Override markers.

CheckerFramework (-PenableCheckerframework=true):
- CodecContext / PgArray: castNonNull at the public getter and
  applyTypeMapping boundary so the nullable backing fields don't
  leak into the API.
- CodecContext.getMappedClass and withTypeMap reject the
  test-only context shape explicitly instead of dereferencing a
  nullable connection/codecs/javaTypes/encoding.
- CodecContext.getTimestampUtils now defers to getConnection() to
  reuse the same null guard.
- CodecRegistry / JavaTypeRegistry / PGRange constructors:
  @SuppressWarnings({"this-escape", "method.invocation"}) for the
  initialize-helpers invoked from the ctor.
- CodecDepth: @SuppressWarnings("type.argument") for the
  ThreadLocal<Integer> withInitial sample-by-supplier pattern.
- CompositeCodec encode/decode: tighten @Nullable on the attribute
  arrays (Object[] → @Nullable Object[]), and tighten
  encodeBinaryFields' fieldData to byte[] @Nullable [].
- CompositeCodec.parseCompositeText returns @Nullable String[].
- IntervalCodec.decodeAsString / JsonCodec / JsonbCodec /
  XmlCodec.decodeAsString: declare @Nullable on the array-byte
  overload (matches the interface default).
- PgSQLInput now declares its attributeValues as @Nullable
  BufferType[] (per-element nullability) and is annotated with
  @SuppressWarnings({"override.return", "override.param"}) to bridge
  the JDBC stub mismatch.
- PgSQLInput/PgSQLOutput constructors fall back to
  ctx.getTypeInfo().getFields(oid) when PgType.getFields() is null.
- PgSQLInputBinary/Text and PgSQLOutputBinary/Text: cacheCodecs uses
  the (oid, pgType) overload so dynamic OIDs route through the
  CompositeCodec / typtype dispatch.
- PgConnection.getObject(typeName,...) returns from decodeBinaryAs/
  decodeTextAs only when the codec actually produced a value;
  otherwise falls through to the PGobject path (the method is
  declared non-null).
- ObjectName.equals(@Nullable Object).
- PgPreparedStatement.setCharacterStream: catch UnsupportedCharsetException
  instead of UnsupportedEncodingException (the throw site uses
  Charset.forName).
- PgResultSet.decodeAsFloat/Double/BigDecimal text branches:
  castNonNull around getFixedString since the codec API expects
  non-null.
- PgStruct.getAttributes(): declare the @Nullable Object[] semantics
  (matches the SQL NULL contract).
- PGUnknownBinary.toString: hoist the @Nullable byte[] field into a
  local to satisfy iteration nullability.
- RangeCodec.decodeText: tolerate elementOid == 0 (pg_range subtype
  not loaded yet); RangeCodec.encodeBinary: castNonNull on the
  range bounds.
- TypeInfoCache: castNonNull around the nullable ResultSet getString
  / getTypeForAlias entry points.

Checkstyle (autostyleApply + manual fixes):
- Add the standard /* Copyright */ header to ObjectName / PgType /
  PgArrayType / PgField (already had it) / TypeInfoCacheLoader, and
  to the new TypeInfoCacheTest / CreateDropStructTest test files.
- PgType: re-indent the three switch blocks to the project's case-
  at-column-6 style; add default branches where Checkstyle's
  fall-through rule required it.
- PgPreparedStatement / PgParameterMetaData: drop the extra blank
  line between package and the first import (and the duplicate
  blank line before the constructor).
- CodecRegistry / BinaryCodec javadoc: insert the blank line that
  Checkstyle expects before <p>.
- TypeInfoCache: ModifierOrderCheck (final static → static final),
  and drop a commented-out body that the indentation check flagged.
- StructTest: switch the multi-line string-concat to lead with '+'
  on the new line (project OperatorWrap convention).
- TypeInfoCacheTest: 'try (' / 'while (' whitespace.
- JsonCodecTest: replace unicode escapes with the literal characters
  to satisfy AvoidEscapedUnicodeCharacters.
- GetObject310Test: align the commented-out @Test markers with
  CommentsIndentationCheck.
- pgjdbc/build.gradle.kts: drop the trailing comma in the listOf(...)
  passed to relocate so autostyleKotlinGradleProcess can parse it.

`./gradlew :postgresql:styleCheck :postgresql:jandex
   -PenableErrorprone -PenableCheckerframework=true` now passes.

+++ pgjdbc/src/main/java/org/postgresql/jdbc/codec/Int8Codec.java
@@ -179,10 +179,10 @@ public final class Int8Codec implements BinaryCodec, TextCodec {
       return (T) Byte.valueOf((byte) value);
     }
     if (targetClass == Double.class) {
-      return (T) Double.valueOf(value);
+      return (T) Double.valueOf((double) value);
     }
     if (targetClass == Float.class) {
-      return (T) Float.valueOf(value);
+      return (T) Float.valueOf((float) value);
     }
     if (targetClass == BigDecimal.class) {
       return (T) BigDecimal.valueOf(value);
